feat: dowith task success 조회 api - presentation 레이어 통합
