+++ src/main/java/com/LetMeDoWith/LetMeDoWith/presentation/task/dto/GenerateDowithTaskSuccessImageUploadPresignedUrlsResDto.java
@@ -2,4 +2,5 @@ package com.LetMeDoWith.LetMeDoWith.presentation.task.dto;
 
 import java.util.List;
 
-public record GenerateDowithTaskSuccessImageUploadPresignedUrlsResDto(List<String> presignedUrls, String method) {}
+public record GenerateDowithTaskSuccessImageUploadPresignedUrlsResDto(
+        List<String> publicImageUrls, List<String> presignedUrls, String method) {}
